+++ src/main/java/com/snnsoluciones/backnathbitpos/dto/orden/ItemCocinaResponse.java
@@ -1,12 +1,13 @@
 package com.snnsoluciones.backnathbitpos.dto.orden;
 
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 import java.util.List;
 
 public record ItemCocinaResponse(
     Long id,
     String productoNombre,
-    Integer cantidad,
+    BigDecimal cantidad,
     String notas,
     List<String> opciones,
     LocalDateTime fechaEnvioCocina,
