+++ src/main/java/com/snnsoluciones/backnathbitpos/dto/orden/OpcionCompuestaRequest.java
@@ -1,9 +1,10 @@
 package com.snnsoluciones.backnathbitpos.dto.orden;
 
 import jakarta.validation.constraints.NotNull;
+import java.math.BigDecimal;
 
 public record OpcionCompuestaRequest(
     @NotNull Long slotId,
     @NotNull Long productoOpcionId,
-    Integer cantidad
+    BigDecimal cantidad
 ) {}
